Fix return of wrong pattern piece
The getPatternIncluding method would return quoted pieces if
the check contents of the quoted part included the format letter.

For instance for Danish (da_DK), the pattern is "EEE 'den' d. MMM"
If we are calling `getPatternIncluding("d", "da_DK")`, the method
would return `"den'"´, but we would like to return `"d."`.

This commit ignores pieces with apostrophe ("'") in them.